machine_signal:State: calc dist with sqrtf

+++ src/plugins/machine_signal/state.cpp
@@ -59,7 +59,7 @@ char const *SignalState::get_debug_G() {
 float SignalState::distance(std::list<signal_rois_t_>::iterator const &s) {
   int dx = s->yellow_roi->start.x + s->yellow_roi->width/2 - pos.x;
   int dy = s->yellow_roi->start.y + s->yellow_roi->height/2 - pos.y;
-  return (float)sqrt(dx*dx + dy*dy);
+  return sqrtf(float(dx*dx + dy*dy));
 }
